backend/requirements/dev.txt
-r common.txt
django-debug-toolbar==3.8.1